# vioapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/ip_location/{ip}": {
            "get": {
                "security": [
                    {
                        "Token, OAuth2Password": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ip_information"
                ],
                "summary": "Get information about the IP address' location (e.g. country, city)",
                "parameters": [
                    {
                        "type": "string",
                        "example": "220.235.222.173",
                        "description": "IP address",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.IPLocation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Location for given IP - not found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.ErrorMessage": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "server.IPLocation": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "country_code": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "mystery_value": {
                    "type": "integer"
                }
            }
        }
    }
}